+++ backend/src/baserow/contrib/database/api/v0/rows/errors.py
@@ -1,2 +1,8 @@
-ERROR_ROW_DOES_NOT_EXIST = ('ERROR_ROW_DOES_NOT_EXIST', 404,
-                            'The requested row does not exist.')
+from rest_framework.status import HTTP_404_NOT_FOUND
+
+
+ERROR_ROW_DOES_NOT_EXIST = (
+    'ERROR_ROW_DOES_NOT_EXIST',
+    HTTP_404_NOT_FOUND,
+    'The requested row does not exist.'
+)