+++ backend/src/baserow/contrib/database/api/v0/views/grid/errors.py
@@ -1,11 +1,14 @@
+from rest_framework.status import HTTP_404_NOT_FOUND, HTTP_400_BAD_REQUEST
+
+
 ERROR_GRID_DOES_NOT_EXIST = (
     'ERROR_GRID_DOES_NOT_EXIST',
-    404,
+    HTTP_404_NOT_FOUND,
     'The requested grid view does not exist.'
 )
 
 ERROR_UNRELATED_FIELD = (
     'ERROR_UNRELATED_FIELD',
-    400,
+    HTTP_400_BAD_REQUEST,
     'The field is not related to the provided grid view.'
 )